PP-10240 Merge wrapper.js into base.client.js
Moving the function in wrapper.js into base.client.js should make it
easier to follow how the base client works.
