+++ app/services/clients/base-client/base.client.js
@@ -1,7 +1,15 @@
 'use strict'
 
 const request = require('request')
-const wrapper = require('./wrapper')
+const correlator = require('correlation-id')
+const joinURL = require('url-join')
+const lodash = require('lodash')
+const { CORRELATION_HEADER } = require('../../../utils/correlation-header')
+const requestLogger = require('../../../utils/request-logger')
+const { RESTClientError } = require('../../../errors')
+
+// Constants
+const SUCCESS_CODES = [200, 201, 202, 204, 206]
 
 // Create request.defaults config
 const requestOptions = {
@@ -18,15 +26,6 @@ const requestOptions = {
 
 const client = request.defaults(requestOptions)
 
-// Export base client
-module.exports = {
-  get: wrapper(client, 'get'),
-  post: wrapper(client, 'post'),
-  put: wrapper(client, 'put'),
-  patch: wrapper(client, 'patch'),
-  delete: wrapper(client, 'delete')
-}
-
 /**
  * @param  {Null | Object} err
  * @param  {Object} response
@@ -35,4 +34,69 @@ module.exports = {
  */
 function retryStrategy (err, response, body) {
   return process.env.NODE_ENV === 'production' && (err && ['ECONNRESET'].includes(err.code))
+}
+
+function makeRequest (method, opts) {
+  return new Promise((resolve, reject) => {
+    opts.method = method.toUpperCase()
+
+    const loggingContext = {
+      correlationId: correlator.getId(),
+      startTime: new Date(),
+      url: joinURL(lodash.get(opts, 'baseUrl', ''), opts.url),
+      method: opts.method,
+      description: opts.description,
+      service: opts.service,
+      additionalLoggingFields: opts.additionalLoggingFields
+    }
+
+    lodash.set(opts, `headers.${CORRELATION_HEADER}`, loggingContext.correlationId)
+    opts.headers['Content-Type'] = opts.headers['Content-Type'] || 'application/json'
+
+    // Set up post response and error handling method
+    const transform = opts.transform || false
+    opts.transform = undefined
+
+    // start request
+    requestLogger.logRequestStart(loggingContext)
+    const call = client(opts, (err, response, body) => {
+      if (err) {
+        reject(err)
+      } else if (response && SUCCESS_CODES.includes(response.statusCode)) {
+        // transform our output if the appropriate function was passed.
+        body = transform ? transform(body) : body
+        resolve(body)
+      } else {
+        let errors = body && (body.message || body.errors)
+        if (errors && Array.isArray(errors)) {
+          errors = errors.join(', ')
+        }
+        const message = errors || body || 'Unknown error'
+        const errorIdentifier = body && body.error_identifier
+        const reason = body && body.reason
+        const err = new RESTClientError(message, opts.service, response.statusCode, errorIdentifier, reason)
+        reject(err)
+      }
+    })
+    // Add event listeners for logging
+    call.on('error', err => {
+      requestLogger.logRequestEnd(loggingContext)
+      requestLogger.logRequestError(loggingContext, err)
+    })
+    call.on('response', response => {
+      requestLogger.logRequestEnd(loggingContext, response)
+      if (!(response && SUCCESS_CODES.includes(response.statusCode))) {
+        requestLogger.logRequestFailure(loggingContext, response)
+      }
+    })
+    return call
+  })
+}
+
+module.exports = {
+  get: opts => makeRequest('get', opts),
+  post: opts => makeRequest('post', opts),
+  put: opts => makeRequest('put', opts),
+  patch: opts => makeRequest('patch', opts),
+  delete: opts => makeRequest('delete', opts)
 }
